# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "LendMe API Server"
        }
    ],
    "paths": {
...
        "/send_request_amount/{id}": {
            "post": {
                "tags": [
                    "request"
                ],
                "summary": "send notificatio to request money",
                "description": "send notificatio to request money",
                "operationId": "send notificatio to request money",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "contact id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/send_refused_amountrequest/{id}/{req_id}": {
            "post": {
                "tags": [
                    "request"
                ],
                "summary": "send refused notificatio to request money",
                "description": "send refused notificatio to request money",
                "operationId": "send refused notificatio to request money",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "contact id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "req_id",
                        "in": "path",
                        "description": "request  id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/send_confirm_requestamount/{id}/{req_id}": {
            "post": {
                "tags": [
                    "request"
                ],
                "summary": "send confirm notificatio to request money",
                "description": "send confirm notificatio to request money",
                "operationId": "send confirm notificatio to request money",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "contact id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "req_id",
                        "in": "path",
                        "description": "request  id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}